pyproject.toml
[tool.poetry]
name = "ai-message"
version = "0.1.0"
description = ""
authors = ["Niwaiwa <po2846@gmail.com>"]
readme = "README.md"
packages = [{include = "ai_message"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
line-bot-sdk = "^2.4.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
